This is an example install batch file
You will have to change the libraries it installs and change it from 2.7 to 3.4. Notice there are some libraries that I downloaded the windows whl for, so the batch can directly install them without going online. These are libraries pip has trouble with. You may not run into this case.